fix some warnings

+++ OgreMain/src/OgreCompositorInstance.cpp
@@ -702,14 +702,12 @@ void CompositorInstance::createResources(bool forResizeOnly)
             
             if(width == 0)
             {
-                width = static_cast<size_t>(
-                                            static_cast<float>(mChain->getViewport()->getActualWidth()) * def->widthFactor);
+                width = static_cast<float>(mChain->getViewport()->getActualWidth()) * def->widthFactor;
                 width = width == 0 ? 1 : width;
             }
             if(height == 0)
             {
-                height = static_cast<size_t>(
-                                             static_cast<float>(mChain->getViewport()->getActualHeight()) * def->heightFactor);
+                height = static_cast<float>(mChain->getViewport()->getActualHeight()) * def->heightFactor;
                 height = height == 0 ? 1 : height;
             }
